Define new property qt.install.dir
qt.install.dir property is used as the -prefix value which is
used in qt.configure.args. This allows the value to be overridden
in qt.install.dir if needed (e.g. in shadowbuilds).

Also override qt.install.dir property in shadow-build stage and
remove the qt.configure.args since there is no need to override
it anymore in shadow-build stage.

Change-Id: I2893924604c543c8739219ccf14564222fb80022
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>