Don't use jom by default in Qt5 Windows builds
Due to QTBUG-23376, qmake runs cannot be safely parallelized when
compiling QtDeclarative on Windows (or anything which depends on
QtDeclarative).

The simplest way to work around this problem is to stop using jom
(make-level parallelization) and fall back to nmake with CL=/MP
(compiler-level parallelization), which does not parallelize qmakes.

For most projects, the build performance is roughly comparable.
QtBase should keep using jom, as this repo is not affected by
QTBUG-23376, and jom is considerably faster than nmake for QtBase
due to the large amount of directories containing a small amount of
source files (under tests/auto).

Task-number: QTBUG-23376
Change-Id: I8bc58c7296797657cc8c682271e964cf9a6390c5
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>