Qt5: enable autotests (insignificant) for OSX 10.7 developer-build
This configuration had autotests disabled due to a high crash rate,
which unacceptably slowed CI.  Now we assert that Qt5 has matured
enough for the situation to be improved (or we should now spend the time
to improve it).

Tests will be marked as significant on a module-by-module basis, as per
standard procedure.

Change-Id: Idc4477e28f7967abee39c4addf29b9643060959c
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>