Windows: fixed incorrect -prefix option when testing qt5.git
We had hardcoded $(base.dir)\qt\qtbase, which works for all modules
except qt5.git.  Use $(qt.install.dir) instead, which is consistent with
the configuration for Linux and Mac.

Note that this also implies we switch from using \ directory separators
to / for the -prefix argument.  This ought to be OK (it's arguably a
configure.exe bug otherwise, since Qt generally recommends usage of the
/ separator).

Change-Id: I8e7f7e51f3edfc8e5dfcfc6533c3eb24d024956b
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>