mac: decrease the parallelization of builds further
Try to decrease the chance of hitting the mysterious build errors
described in QTQAINFRA-337 and QTQAINFRA-376, which are theorized to be
more likely when system load is high.

The test mac minis themselves have two cores and use distcc without a
scheduler, so with -j6 it is still possible that the total amount of
compile jobs could outnumber the amount of available cores by a
factor of 3 to 1, if every mac is building at the same time.

Task-number: QTQAINFRA-337
Task-number: QTQAINFRA-376
Change-Id: I2b403f86cac77bbd90d9319964ef8ab0844551e2
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>