Windows: enable non-enforcing autotests for all remaining Qt5 modules
Windows autotests have been enabled for various modules for some time,
and recently were enabled even for QtBase, the largest and most complex
module.  We are reasonably confident that the test scripts are working
correctly and the test runtimes are acceptable, so there is no reason to
leave these disabled any longer.

Change-Id: Ifa838d586ce32fcf89757c79c17715fc44aa2814
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>