Replace -no-xcb -no-wayland with -no-qpa-platform-guard where possible
Prior to the -no-qpa-platform-guard option, the only way to configure
qtbase without enabling one of the QPA platforms hosted in qtbase was to
pass -no-<backend> for each backend.  This meant there was no
stable configure line which would remain working after adding new
backends, blocking the integration of changes such as
http://codereview.qt-project.org/19218

-no-qpa-platform-guard is the new way to solve this problem.

Unfortunately it does not yet exist in all branches, which is why this
change is applied only to those branches where the feature is thought to
exist.

Change-Id: I4a0d583772a357ac02a58d0c2ef296bf9aae52bc
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>