Added a new revdep-qtdeclarative configuration.
Test qtdeclarative master on top of incoming changes to other modules.

This configuration was chosen because it is fast and has the best test
coverage.

Note that, although defined at the "Qt Modules Continuous Integration"
level, this configuration is not valid for all modules.  A corresponding
stage in Pulse should only be set up for those modules where it makes
sense.

Change-Id: I5b4994797fb7732c8fba463e6a33df06ebfac9e8
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>