Restore the "win32-msvc2010 developer-build qtnamespace" config
We should eventually aim to have qtnamespace included in the second
win32-msvc2010 configuration.  However, it was removed because we wanted
to see the results of the autotests for a developer build on Windows -
and -qtnamespace can't yet pass compilation.

For now, let's have _both_ the configurations in testconfig, and the
appropriate configuration can be selected at the Pulse level (or both of
them can be enabled, to compare results).

Change-Id: I32a34db0643fabced5414fe20c83788fd62e1e7b
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>