Qt5 CI: use testscheduler script for running autotests
This is a new method of running autotests built on top of "make check".

The script has been tested with success separately from the main CI for
a while, and seems ready to go into general use.  Currently the benefits
include improved reporting of failures (e.g. the summary will tell
how much time was spent on insignificant tests).

The testscheduler also supports running autotests in parallel if their
.pro file is marked with CONFIG+=parallel_test, but that needs some
additional testing; we need to ensure CONFIG+=parallel_test is used
on autotests correctly, to avoid decreasing the stability of test runs.

Change-Id: I8d6ef7c268988b003367b2aeaf1272535547dc1c
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>
Reviewed-by: Jyri Tahtela <jyri.tahtela@nokia.com>
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>