Set QMAKEPATH for buildsystem branch compatibility
The "configure && make install" process used for this build doesn't
work with the buildsystem branches. Setting QMAKEPATH as a workaround
should allow it to work.

Once the branches are merged, the test script should be modified to use
a "configure && make && make install" process, and this change can be
reverted.

Change-Id: I013c15567d2e70381741abbd26e194678f73a36e
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>