CI: set environment variable to flag environment as Qt Project CI
When there is a need to disable or change a portion of a test
specifically when run in a continuous integration environment (e.g. skip
possibly flaky parts of a test), the QT_TEST_CI environment variable can
be checked.

This should be generally discouraged, used as a last resort or
intermediate measure to make tests useful.

Change-Id: I666dc82b289ed23d5dbc90c6646929b483c3c02c
Reviewed-by: Kurt Korbatits  <kurt.korbatits@nokia.com>
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>