Added 'Qt5 submodule update' project
This project, used to update the qt5.git submodule pointers, was not
ported from the Nokia-internal pulseconfig repository at the same
time as the rest of the projects were - probably by accident due to
it not following the same naming scheme as other projects.

This commit copies the project config exactly from the pulseconfig
repository.

Change-Id: Ie90e7a6fa163a92f25719439db89ea1d6bf26bcd
Reviewed-by: Toby Tomkins <tjtomkins@gmail.com>