jenkins: improved reliability and error reporting of automated reboot
Running the reboot command through a post-build groovy script seems to
involve serializing a lot of data between master and slave, and various
synchronous remote procedure calls, making the process sensitive to
flaky or high-latency network connections.

It could also be interrupted if a build is cancelled during the
post-build step.

Make it more robust by repeatedly attempting the reboot.
This can recover if a network problem occurs or the build is cancelled
during the post-build step.
Reboot problems are posted back to the build as warning badges.

If the reboot ultimately can't succeed, take the node out of
"temporarily offline" state, to prevent the accumulation of nodes in
this state.

Change-Id: Iff69dfb4499d8818da07d0d7d916e9c8601c7fdd
Reviewed-by: Janne Anttila <janne.anttila@digia.com>
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>