jenkins: further improve reliability and reporting of reboot hook
When rebooting succeeds, put an info badge on the build, so there's a
record of the successful reboot.

Use "-n" option to sudo, which makes sudo fail (instead of block) if
passwordless "sudo reboot" is not set up correctly.

When adding a warning badge relating to a reboot failure, include the
entire error stacktrace instead of only the summary.

When rebooting fails, try to shut down and reinitialize the channel
between master and slave before trying the reboot again. This appears to
enable recovery from certain errors which could block the reboot, most
notably "Could not initialize class
org.codehause.groovy.runtime.InvokerHelper" from Jenkins 1.466 when
attempting to invoke a groovy script remotely.

Reduce the maximum amount of reboot attempts, since the expected runtime
of each attempt is now higher.

Change-Id: Ie95aa6ce94cc542f51d3429530ae26aa279bbfe0
Reviewed-by: Toby Tomkins <tjtomkins@gmail.com>
Reviewed-by: Janne Anttila <janne.anttila@digia.com>