Use testscheduler to run tests also for Qt4.
This script has been used for some time to run Qt5 tests, but wasn't
applied to Qt4, mainly because the time was not set aside to make sure
it works correctly.

Now it seems that all tests work fine with the testscheduler in Qt 4.8
as well, so enable it there.

The main benefits are:

  - easier to parse logs (for humans and scripts)
  - summary output at end of test run ("TEST RUN COMPLETED!")
  - opens the way toward using CONFIG+=parallel_test also in Qt4
    (although that is not enabled as part of this commit, as it carries
    its own risk).

Change-Id: Ifad38ce0c33cc20c8759d3e108c2e537c02d79d0
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>