Add Qt5 CI config for Windows 7 with Angle
For now this configs has 'forcesuccess' and 'qt.tests.insignificant'
enabled. The idea is that we would run this configuration in parallel
with existing one at least for a while. When all CI projects are
passsing with Angle enabled build we can probably use Angle with one
of the existing windows configurations and possibly remove this one.

Change-Id: I5e8cf212a4d65f5a57153971426b3b609fdae373
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>