TestConfig: Initial version for WEC7 ARM CI builds.
Added new build only CI stage to build for WEC7. The stage is enabled
for almost all CI repos and branches (except those stages where
configurations are overridden).

The stage has forcesuccess enabled globally. It remains to be seen if
this stage will ever be made enforcing; the main reason for this stage
right now is that there is some place from where community can check the
status of WEC7 build.

Cleaned up LIB/INCLUDE variables based on stage name before calling
vsvars32.bat and setting environment for WEC7. The reason for this is
that LIB/INCLUDE contain desktop icu/openssl headers and libraries,
and thus are not compatible with WEC7.

Change-Id: Ib857ae8f580890bc3e5d76478b1e63a5f308996e
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>