Use jom for all Qt5 CI jobs.
Based on test on dev-ci jom 1.0.13 seems to be stable and is able
to build different Qt5 CI project successfully => Take it in use
for all Qt5 CI jobs.

In addition remove hard coded -j options, since jom 1.0.13 defaults
to number of cores in machine and number of cores in Windows CI VMs
can differ.

Change-Id: I72020d37962273a65d626b330792b0a3cfd611b4
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>