Push down forcesuccess and qt.tests.insignificant properties for WEC7.
In order to make it possible to remove the forcesuccess and
qt.tests.insignificant properties on repo and branch basis,
push down the forcesuccess and qt.tests.insignificant properties.

The WEC7 CI stage is at the moment build only, but added
qt.tests.insignificant properties at the same time with forcesuccess.

Change-Id: I44906f0e5d969e92294b1b51cb67799ecf3defb2
Reviewed-by: Simo Fält <simo.falt@digia.com>
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>