Testconfig: Force QtWayland to succeed
While there is an error in jenkins ini file, we are missing 10
OpenSuse 12.1 nodes from CI. To fix that, we first have rename
those nodes which have Wayland manually compiled to system. But
before doing that we have to add those nodes to same range with
rest of the OpenSuse 12.1 nodes to be able to add nodes which can
be used, when those 5 nodes having Wayland installed are taken
down and renamed. After having all OpenSuse 12.1 nodes within same
range, we can't quarantee that QtWayland is built on such node that
have the Wayland, therefore I'm forcing those to succeed for now.

Change-Id: I571529f3569bcf3f84e21e38467125b00f37a82a
Reviewed-by: Andrew Knight <andrew.knight@digia.com>
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>