Add Windows 8.1 x64 builds to the CI
This modifies Jenkin's environment setup (ICU path,
lib and include) so that current Win8 machines
have to be modified before deploying these settings.

This also removes the setting of ICU4C paths in
current Win 8 machines via Jenkins, as it will be
replaced by the proper ICU version and fixed paths
in variables.

Change-Id: I7db0685835e46e85e1375312e85f3b848c9afb0e
Reviewed-by: Simo Fält <simo.falt@digia.com>