Stop configuring with -debug-and-release on Linux
This doesn't really work, we don't have a D suffix like on Windows
or _debug on Mac OS X for debug builds. Without a suffix a debug
and release build library gets the same name and they're overwritten
or - as it's a race condition - neither of them will end up where
they should.

Change-Id: Ifbd383b0c0b61c656ab45ad904271232edc7451c
Reviewed-by: Zeno Albisser <zeno.albisser@digia.com>
Reviewed-by: Jani Heikkinen <jani.heikkinen@digia.com>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@digia.com>