Prospective fix for QtQA master integrations
The integration performs a build of qtbase, which fails on Windows because
qt5.git is checked out from the 5.4 branch, which has outdated win_bison
binaries. So the fix is simple: If we build qtbase from dev, then we must also
use qt5 dev, to pull in the correct gnuwin32 binaries.

Change-Id: I2c982ad412da84f15e789f8785b14ef7c59cc2bd
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>