Mark qtlocation and qtsensors 5.4 integration on OpenSUSE as insignificant
This is the only 5.4.x module that had opensuse as significant config, but
unfortunately it doesn't work. The cmake installation is too old.

We decided to bite the bullet and ignore this cmake business in 5.4, given
that 5.4.2 might be the last 5.4 release.

Instead in 5.5 we're going to try to do the right thing:

1) Make Ubuntu 14.04 enforced in the CI (blacklist tst_qwindow)
2) Make sure cmake is up-to-date there and that the tests pass
3) Uninstall cmake from OpenSuSE and then also make that enforcing

Change-Id: I88fa722129d3955939389dfe674c237d83b9792e
Reviewed-by: Tony Sarajärvi <tony.sarajarvi@digia.com>